+++ src/rearrange.cpp
@@ -9,7 +9,7 @@ typedef int_fast16_t int16;
 const int16 UNDEFINED = -1;
 
 // Assumptions: 
-//  * Tree is bifurcating and rooted on a tip.
+//  * Tree is bifurcating and rooted on a tip; root node is labelled with n_tip + 1
 //  [[Rcpp::export]]
 IntegerMatrix nni(const IntegerMatrix edge,
                   const IntegerVector randomEdge,
@@ -20,9 +20,9 @@ IntegerMatrix nni(const IntegerMatrix edge,
     n_tip = (n_edge / 2) + 1;
     
   int16 n_samplable = 0;
-  int16 *samplable = new int16[n_edge];
+  std::unique_ptr<int16[]> samplable = std::make_unique<int16[]>(n_edge);
   for (int16 i = n_edge; --i; ) {
-    if (edge(i, 1) > n_tip) {
+    if (edge(i, 1) > n_tip && edge(i, 0) != n_tip + 1) {
       samplable[n_samplable++] = i;
     }
   }
@@ -34,7 +34,6 @@ IntegerMatrix nni(const IntegerMatrix edge,
     edge_to_break = samplable[chosen_edge % n_samplable],
     end1 = edge(edge_to_break, 0),
     end2 = edge(edge_to_break, 1);
-  delete[] samplable;
   int16
     ind1 = UNDEFINED,
     ind2 = UNDEFINED;
@@ -78,10 +77,13 @@ IntegerMatrix spr_moves(const IntegerMatrix edge) {
   if (edge(0, 0) != root_node) throw std::invalid_argument("edge[1,] must connect root to leaf. Try Preorder(root(tree)).");
   if (edge(1, 0) != root_node) throw std::invalid_argument("edge[2,] must connect root to leaf. Try Preorder(root(tree)).");
   
-  int16* prune = new int16[(n_edge - 1) * (n_edge - 3)];
-  int16* graft = new int16[(n_edge - 1) * (n_edge - 3)];
-  int16* above = new int16[(n_edge - 1) * (n_edge - 3)];
-  int16* bside = new int16[(n_edge - 1) * (n_edge - 3)];
+  
+  std::unique_ptr<int16[]> 
+    prune = std::make_unique<int16[]>((n_edge - 1) * (n_edge - 3)),
+    graft = std::make_unique<int16[]>((n_edge - 1) * (n_edge - 3)),
+    above = std::make_unique<int16[]>((n_edge - 1) * (n_edge - 3)),
+    bside = std::make_unique<int16[]>((n_edge - 1) * (n_edge - 3))
+  ;
   int16 n_moves = 0, root_daughter_2 = 0;
   
   // Root edge first
@@ -159,7 +161,6 @@ IntegerMatrix spr_moves(const IntegerMatrix edge) {
     }
   }
   
-  
   IntegerMatrix ret(n_moves, 4);
   for (int16 i = n_moves; i--; ) {
     ret(i, 0) = prune[i];
@@ -167,10 +168,6 @@ IntegerMatrix spr_moves(const IntegerMatrix edge) {
     ret(i, 2) = above[i];
     ret(i, 3) = bside[i];
   }
-  delete[] prune;
-  delete[] graft;
-  delete[] above;
-  delete[] bside;
   return (ret);
 }
 
@@ -239,14 +236,15 @@ IntegerMatrix tbr_moves(const IntegerMatrix edge) {
   if (edge(0, 0) != root_node) throw std::invalid_argument("edge[1,] must connect root to leaf. Try Preorder(root(tree)).");
   if (edge(1, 0) != root_node) throw std::invalid_argument("edge[2,] must connect root to leaf. Try Preorder(root(tree)).");
   
-  std::unique_ptr<int16[]> n_edges_above = std::make_unique<int16[]>(n_edge);
-  std::unique_ptr<int16[]> probibited_parent = std::make_unique<int16[]>(n_edge);
-  std::unique_ptr<int16[]> probibited_sibling = std::make_unique<int16[]>(n_edge);
-  
-  int16* prune = new int16[(n_edge - 1) * (n_edge - 3)];
-  int16* graft = new int16[(n_edge - 1) * (n_edge - 3)];
-  int16* above = new int16[(n_edge - 1) * (n_edge - 3)];
-  int16* bside = new int16[(n_edge - 1) * (n_edge - 3)];
+  std::unique_ptr<int16[]>
+    n_edges_above = std::make_unique<int16[]>(n_edge),
+    probibited_parent = std::make_unique<int16[]>(n_edge),
+    probibited_sibling = std::make_unique<int16[]>(n_edge),
+    prune = std::make_unique<int16[]>((n_edge - 1) * (n_edge - 3)),
+    graft = std::make_unique<int16[]>((n_edge - 1) * (n_edge - 3)),
+    above = std::make_unique<int16[]>((n_edge - 1) * (n_edge - 3)),
+    bside = std::make_unique<int16[]>((n_edge - 1) * (n_edge - 3))
+  ;
   int16 n_moves = 0, root_daughter_2 = 0;
   
   // Root edge first
@@ -278,10 +276,6 @@ IntegerMatrix tbr_moves(const IntegerMatrix edge) {
     ret(i, 2) = above[i];
     ret(i, 3) = bside[i];
   }
-  delete[] prune;
-  delete[] graft;
-  delete[] above;
-  delete[] bside;
   return (ret);
 }