+++ NEWS.md
@@ -1,6 +1,7 @@
 # TreeSearch 0.4.3.9001 (development)
 
  - 'shiny' graphical user interface for tree search.
+ - Remove redundant function `AsBinary()`.
  - C implementations of rearrangement functions.
 
 # TreeSearch 0.4.3