+++ NEWS.md
@@ -1,6 +1,9 @@
 # TreeSearch 0.4.3.9001 (development)
 
  - 'shiny' graphical user interface for tree search.
+ - `Consistency()` calculates consistency and retention 'indices'.
+ - `MinimumLength()` and `MaximumLength()` calculate range of possible lengths
+   of characters in a dataset on any tree.
  - Remove redundant function `AsBinary()`.
  - C implementations of rearrangement functions.