+++ NEWS.md
@@ -2,6 +2,7 @@
 
  - 'shiny' graphical user interface for tree search.
  - Remove redundant function `AsBinary()`.
+ - Merge `Fitch()` / `IWScore()` to combined function `TreeLength()`
  - C implementations of rearrangement functions.
 
 # TreeSearch 0.4.3