+++ NEWS.md
@@ -1,8 +1,10 @@
 # TreeSearch 0.4.3.9001 (development)
 
- - 'shiny' graphical user interface for tree search.
- - Remove redundant function `AsBinary()`.
- - C implementations of rearrangement functions.
+ - 'shiny' graphical user interface for tree search
+ - Support for topological constraints during tree search
+ - Remove redundant function `AsBinary()`
+ - Merge `Fitch()` / `IWScore()` to combined function `TreeLength()`.
+ - C implementations of rearrangement functions
 
 # TreeSearch 0.4.3