+++ NEWS.md
@@ -4,6 +4,7 @@
   - `Consistency()` calculates consistency and retention 'indices'
  - `MinimumLength()` and `MaximumLength()` calculate range of possible lengths
    of characters in a dataset on any tree
+ - Label nodes with split frequencies using `JackLabels(plot = FALSE)`
  - Support for topological constraints during tree search
  - Remove redundant function `AsBinary()`
  - C implementations of rearrangement functions